refactor: move (un)sew/link methods to dedicated module
